+++ Source/ACE/Entity/CollidableObject.cs
@@ -1,17 +1,17 @@
-﻿namespace ACE.Entity
-{
-    public abstract class CollidableObject : WorldObject
-    {
-        internal CollidableObject(AceObject aceO)
-            : base(aceO)
-        {
-        }
+﻿////namespace ACE.Entity
+////{
+////    public abstract class CollidableObject : WorldObject
+////    {
+////        internal CollidableObject(AceObject aceO)
+////            : base(aceO)
+////        {
+////        }
 
-        internal CollidableObject(ObjectGuid guid, AceObject aceO)
-            : base(guid, aceO)
-        {
-        }
+////        internal CollidableObject(ObjectGuid guid, AceObject aceO)
+////            : base(guid, aceO)
+////        {
+////        }
 
-        public abstract void OnCollide(ObjectGuid playerId);
-    }
-}
+////        public abstract void OnCollide(ObjectGuid playerId);
+////    }
+////}