+++ Source/ACE/Entity/DebugObject.cs
@@ -1,23 +1,23 @@
-﻿using ACE.Managers;
+﻿////using ACE.Managers;
 
-namespace ACE.Entity
-{
-    public class DebugObject : CollidableObject
-    {
-        public DebugObject(AceObject aceObject)
-            : base(aceObject)
-        {
-        }
+////namespace ACE.Entity
+////{
+////    public class DebugObject : CollidableObject
+////    {
+////        public DebugObject(AceObject aceObject)
+////            : base(aceObject)
+////        {
+////        }
 
-        public DebugObject(ObjectGuid guid, AceObject aceObject)
-            : base(guid, aceObject)
-        {
-            aceObject.AceObjectId = guid.Full;
-        }
+////        public DebugObject(ObjectGuid guid, AceObject aceObject)
+////            : base(guid, aceObject)
+////        {
+////            aceObject.AceObjectId = guid.Full;
+////        }
 
-        public override void OnCollide(ObjectGuid playerId)
-        {
-            // TODO: Implement
-        }
-    }
-}
+////        public override void OnCollide(ObjectGuid playerId)
+////        {
+////            // TODO: Implement
+////        }
+////    }
+////}