+++ Source/ACE/Entity/SummonedPortal.cs
@@ -1,16 +1,16 @@
-﻿using ACE.Entity.Enum;
+﻿////using ACE.Entity.Enum;
 
-namespace ACE.Entity
-{
-    public class SummonedPortal : UsableObject
-    {
-        public SummonedPortal(ItemType type, ObjectGuid guid) : base(type, guid)
-        {
-        }
+////namespace ACE.Entity
+////{
+////    public class SummonedPortal : WorldObject
+////    {
+////        public SummonedPortal(ItemType type, ObjectGuid guid) : base(type, guid)
+////        {
+////        }
 
-        public override void OnUse(ObjectGuid playerId)
-        {
-            // TODO: Implement
-        }
-    }
-}
+////        public override void OnUse(ObjectGuid playerId)
+////        {
+////            // TODO: Implement
+////        }
+////    }
+////}